pyproject.toml
[project]
name = "llm-local"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "google-generativeai>=0.8.5",
    "pandas>=2.2.3",
    "pymupdf>=1.25.5",
    "python-docx>=1.1.2",
    "sentence-transformers>=4.1.0",
]
